refactor: rename package to django_migrate_sql
BREAKING CHANGE: rename package from `migrate_sql` to `django_migrate_sql` to follow Django reusable apps best practices
