+++ README.md
@@ -83,17 +83,17 @@ Install from PyPi:
 pip install django-migrate-sql-deux
 ```
 
-Add `migrate_sql` to `INSTALLED_APPS`:
+Add `django_migrate_sql` to `INSTALLED_APPS`:
 
 ```python
 INSTALLED_APPS = [
     # ...
-    'migrate_sql',
+    'django_migrate_sql',
 ]
 ```
 
 App defines a custom `makemigrations` command, that inherits from
-Django's core one, so in order `migrate_sql` app to kick in put it
+Django's core one, so in order `django_migrate_sql` app to kick in put it
 after any other apps that redefine `makemigrations` command too.
 
 ## Usage
@@ -104,22 +104,22 @@ after any other apps that redefine `makemigrations` command too.
 
 ```python
 # PostgreSQL example.
-# Let's define a simple function and let `migrate_sql` manage it's changes.
+# Let's define a simple function and let `django_migrate_sql` manage its changes.
 
-from migrate_sql.config import SQLItem
+from django_migrate_sql.config import SQLItem
 
 sql_items = [
     SQLItem(
         'make_sum',   # name of the item
         'create or replace function make_sum(a int, b int) returns int as $$ '
-        'begin return a + b; end; ' 
+        'begin return a + b; end; '
         '$$ language plpgsql;',  # forward sql
         reverse_sql='drop function make_sum(int, int);',  # sql for removal
     ),
 ]
 ```
 
-3)  Create migration `./manage.py makemigrations`:
+3)  Create migration `python manage.py makemigrations`:
 
         Migrations for 'app_name':
           0002_auto_xxxx.py:
@@ -129,7 +129,7 @@ You can take a look at content this generated:
 
 ```python
 from django.db import migrations, models
-import migrate_sql.operations
+import django_migrate_sql.operations
 
 
 class Migration(migrations.Migration):
@@ -137,7 +137,7 @@ class Migration(migrations.Migration):
         ('app_name', '0001_initial'),
     ]
     operations = [
-        migrate_sql.operations.CreateSQL(
+        django_migrate_sql.operations.CreateSQL(
             name='make_sum',
             sql='create or replace function make_sum(a int, b int) returns int as $$ begin return a + b; end; $$ language plpgsql;',
             reverse_sql='drop function make_sum(int, int);',
@@ -145,7 +145,7 @@ class Migration(migrations.Migration):
     ]
 ```
 
-4)  Execute migration `./manage.py migrate`:
+4)  Execute migration `python manage.py migrate`:
 
         Operations to perform:
           Apply all migrations: app_name
@@ -153,10 +153,10 @@ class Migration(migrations.Migration):
           Rendering model states... DONE
           Applying app_name.0002_xxxx... OK
 
-Check result in `./manage.py dbshell`:
+Check result in `python manage.py dbshell`:
 
     db_name=# select make_sum(12, 15);
-     make_sum 
+     make_sum
     ----------
            27
     (1 row)
@@ -166,11 +166,11 @@ takes a custom type as argument:
 
 5)  Edit your `sql_config.py`:
 
-``` python
+```python
 # PostgreSQL example #2.
 # Function and custom type.
 
-from migrate_sql.config import SQLItem
+from django_migrate_sql.config import SQLItem
 
 sql_items = [
     SQLItem(
@@ -191,7 +191,7 @@ sql_items = [
 ]
 ```
 
-6)  Generate migration `./manage.py makemigrations`:
+6)  Generate migration `python manage.py makemigrations`:
 
 ```
 Migrations for 'app_name':
@@ -206,7 +206,7 @@ You can take a look at the content this generated:
 
 ```python
 from django.db import migrations, models
-import migrate_sql.operations
+import django_migrate_sql.operations
 
 
 class Migration(migrations.Migration):
@@ -214,24 +214,24 @@ class Migration(migrations.Migration):
         ('app_name', '0002_xxxx'),
     ]
     operations = [
-        migrate_sql.operations.ReverseAlterSQL(
+        django_migrate_sql.operations.ReverseAlterSQL(
             name='make_sum',
             sql='drop function make_sum(int, int);',
             reverse_sql='create or replace function make_sum(a int, b int) returns int as $$ begin return a + b; end; $$ language plpgsql;',
         ),
-        migrate_sql.operations.CreateSQL(
+        django_migrate_sql.operations.CreateSQL(
             name='mynum',
             sql='create type mynum as (num int, name varchar(20));',
             reverse_sql='drop type mynum;',
         ),
-        migrate_sql.operations.AlterSQL(
+        django_migrate_sql.operations.AlterSQL(
             name='make_sum',
             sql='create or replace function make_sum(a mynum, b mynum) returns mynum as $$ begin return (a.num + b.num, \'result\')::mynum; end; $$ language plpgsql;',
             reverse_sql='drop function make_sum(mynum, mynum);',
         ),
-        migrate_sql.operations.AlterSQLState(
+        django_migrate_sql.operations.AlterSQLState(
             name='make_sum',
-            add_dependencies=(('app_name', 'mynum'),),
+            add_dependencies=[('app_name', 'mynum')],
         ),
     ]
 ```
@@ -244,7 +244,7 @@ another version of it, so `DROP` makes it clean.
 definition, it will NOT drop + create it, but just rerun forward SQL,
 which is `CREATE OR REPLACE` in this example.**
 
-7)  Execute migration `./manage.py migrate`:
+7)  Execute migration `python manage.py migrate`:
 
 
 ```
@@ -259,7 +259,7 @@ Check results:
 
 ```
 db_name=# select make_sum((5, 'a')::mynum, (3, 'b')::mynum);
-  make_sum  
+  make_sum
 ------------
  (8,result)
 (1 row)
