+++ tests/test_app/apps.py
@@ -5,5 +5,5 @@ from django.apps import AppConfig
 
 
 class TestAppConfig(AppConfig):
-    name = 'test_app'
+    name = 'tests.test_app'
     verbose_name = 'Test App'
