+++ tests/test_app/migrations_deps_delete/0003_auto_20160108_0048.py
@@ -2,7 +2,7 @@
 from __future__ import unicode_literals
 
 from django.db import models, migrations
-import migrate_sql.operations
+import django_migrate_sql.operations
 
 
 class Migration(migrations.Migration):
@@ -12,17 +12,17 @@ class Migration(migrations.Migration):
     ]
 
     operations = [
-        migrate_sql.operations.ReverseAlterSQL(
+        django_migrate_sql.operations.ReverseAlterSQL(
             name='narration',
             sql='DROP TYPE narration',
             reverse_sql='CREATE TYPE narration AS (sale1 sale, book1 book, arg1 int); -- 1',
         ),
-        migrate_sql.operations.CreateSQL(
+        django_migrate_sql.operations.CreateSQL(
             name='edition',
             sql='CREATE TYPE edition AS (arg1 int); -- 1',
             reverse_sql='DROP TYPE edition',
         ),
-        migrate_sql.operations.ReverseAlterSQL(
+        django_migrate_sql.operations.ReverseAlterSQL(
             name='book',
             sql='DROP TYPE book',
             reverse_sql='CREATE TYPE book AS (arg1 int); -- 1',
