+++ tests/test_app/migrations_deps_update/0002_auto_20160106_0947.py
@@ -2,7 +2,7 @@
 from __future__ import unicode_literals
 
 from django.db import models, migrations
-import migrate_sql.operations
+import django_migrate_sql.operations
 
 
 class Migration(migrations.Migration):
@@ -13,17 +13,17 @@ class Migration(migrations.Migration):
     ]
 
     operations = [
-        migrate_sql.operations.CreateSQL(
+        django_migrate_sql.operations.CreateSQL(
             name='book',
             sql='CREATE TYPE book AS (arg1 int); -- 1',
             reverse_sql='DROP TYPE book',
         ),
-        migrate_sql.operations.CreateSQL(
+        django_migrate_sql.operations.CreateSQL(
             name='rating',
             sql='CREATE TYPE rating AS (arg1 int); -- 1',
             reverse_sql='DROP TYPE rating',
         ),
-        migrate_sql.operations.CreateSQL(
+        django_migrate_sql.operations.CreateSQL(
             name='narration',
             sql='CREATE TYPE narration AS (sale1 sale, book1 book, arg1 int); -- 1',
             reverse_sql='DROP TYPE narration',
