+++ tests/test_app/test_utils.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 from django.test import TestCase
 
-from migrate_sql.autodetector import is_sql_equal
+from django_migrate_sql.autodetector import is_sql_equal
 
 
 class SQLComparisonTestCase(TestCase):
