--- tests/test_app/urls.py
@@ -1,6 +1,0 @@
-from django.urls.conf import path
-from django.contrib import admin
-
-urlpatterns = [
-    path('admin/', admin.site.urls),
-]
