+++ tests/test_app2/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 
 class TestApp2Config(AppConfig):
-    name = 'test_app2'
+    name = 'tests.test_app2'
     verbose_name = 'Test App2'
