+++ tests/test_app2/migrations_deps_delete/0001_initial.py
@@ -2,7 +2,7 @@
 from __future__ import unicode_literals
 
 from django.db import models, migrations
-import migrate_sql.operations
+import django_migrate_sql.operations
 
 
 class Migration(migrations.Migration):
@@ -11,7 +11,7 @@ class Migration(migrations.Migration):
     ]
 
     operations = [
-        migrate_sql.operations.CreateSQL(
+        django_migrate_sql.operations.CreateSQL(
             name='sale',
             sql='CREATE TYPE sale AS (arg1 int); -- 1',
             reverse_sql='DROP TYPE sale',
