--- tests/test_app2/urls.py
@@ -1,10 +1,0 @@
-from django.conf.urls import url
-from django.contrib import admin
-
-urlpatterns = [
-    # Examples:
-    # url(r'^$', 'test_project.views.home', name='home'),
-    # url(r'^blog/', include('blog.urls')),
-
-    url(r'^admin/', admin.site.urls),
-]
